remove some warning - seem to need geometry module for tar ball...
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/branches/release-1.3@7226 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/space/dofmapper/code.hh
@@ -111,13 +111,13 @@ namespace Dune
 
       const unsigned int &operator[] ( unsigned int i ) const
       {
-        assert( i < end() - begin() );
+        assert( i < (unsigned int)(end() - begin()) );
         return begin()[ i ];
       }
 
       unsigned int &operator[] ( unsigned int i )
       {
-        assert( i < end() - begin() );
+        assert( i < (unsigned int)(end() - begin()) );
         return begin()[ i ];
       }
     };
